NFC various fixes (#3134)
* nfc: add furi_hal_nfc_event_stop()

* nfc: fix race condition

* nfc felica: more checks for polling command

* st25r3916: don't read fifo if 0 bytes received

* nfc app: enable skip dict attack when tag is lost

* nfc hal: start event before waiting for specific irq

* iso15 signal: remove incorrect assert

* nfc: fix slix detection

* nfc: check parity bits before copying them

* nfc app: add log size during emulation

* nfc: add iso13239 crc to API

* nfc app: fix listener stop in detect reader

* nfc app: fix PVS warnings

* nfc hal: add event stop documentation

* mf classic: fix var overflow

+++ lib/nfc/protocols/mf_classic/mf_classic_poller_i.h
@@ -55,7 +55,7 @@ typedef enum {
 typedef struct {
     uint8_t current_sector;
     MfClassicSectorTrailer sec_tr;
-    uint8_t current_block;
+    uint16_t current_block;
     bool is_value_block;
     MfClassicKeyType key_type_read;
     MfClassicKeyType key_type_write;
@@ -68,7 +68,7 @@ typedef struct {
     MfClassicKey current_key;
     MfClassicKeyType current_key_type;
     bool auth_passed;
-    uint8_t current_block;
+    uint16_t current_block;
     uint8_t reuse_key_sector;
 } MfClassicPollerDictAttackContext;
